DSP-23094 Automatically creates a product version tag when merging PR with new release changes (#83)
* DSP-23094 Automates product tag creation

Automatically tags repository with product and version when a PR with
new release information is merged.

Adds GitHub action that reads information from a PR title, matches rule
used for new releases (e.g. "DSE 6.8.31 Release"), and automatically
creates product-version tag (e.g. "dse-6.8.31") when the PR is merged.

We have customers that use this repository tags to be automatically
notified for new versions. For that purpose we have introduced in our
release process to manually add a tag with product version when merging
the release notes PR changes.
This GitHub action eliminates the need to manually add the tag.

* DSP-23094 Adds PR template with info about auto tag creation

Adds pull-request template with information about which conditions will
trigger the GitHub Action that will automatically create a product
release tag.